requirements.txt
# Core dependencies
apscheduler==3.10.4
dateparser
faiss-cpu==1.8.0.post1
loguru==0.7.2
openai #==1.107.2
openai-whisper==20231117 # STT (cloud/offline with whisper.cpp alt)
playsound==1.3.0
# pvporcupine==1.9.5
psutil
pvporcupine==3.0.5
pydantic==2.11.9
# pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.3
sounddevice==0.5.2
soundfile==0.13.1
spacy==3.7.5
sqlite-utils #==3.37
TTS==0.22.0

# Embeddings (local-first, optional)
numpy==1.26.4
sentence-transformers==3.0.1

# ElevenLabs (optional voice)
elevenlabs==2.15.1

# Home Assistant REST (optional)
httpx==0.28.1

# Google API for Calendar integration
google-auth-httplib2
google-auth-oauthlib
google-api-python-client

# Google API for Gmail integration
google-auth-httplib2
google-auth-oauthlib
google-api-python-client

# # Optional future
# weaviate-client==3.26.0
# pymilvus==2.4.4